+++ .metals/metals.lock.db
@@ -1,0 +1,6 @@
+#FileLock
+#Fri Oct 21 14:12:57 CST 2022
+server=localhost\:40269
+hostName=localhost
+method=file
+id=183f92c5bf49f17491becab8aaffdbd093a600b4d36
