+++ src/main/scala/nutcore/backend/BackendCommons.scala
@@ -1,0 +1,43 @@
+/**************************************************************************************
+* Copyright (c) 2020 Institute of Computing Technology, CAS
+* Copyright (c) 2020 University of Chinese Academy of Sciences
+* 
+* NutShell is licensed under Mulan PSL v2.
+* You can use this software according to the terms and conditions of the Mulan PSL v2. 
+* You may obtain a copy of Mulan PSL v2 at:
+*             http://license.coscl.org.cn/MulanPSL2 
+* 
+* THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER 
+* EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY OR 
+* FIT FOR A PARTICULAR PURPOSE.  
+*
+* See the Mulan PSL v2 for more details.  
+***************************************************************************************/
+
+package nutcore.backend
+
+import chisel3._
+import chisel3.util._
+import chisel3.util.experimental.BoringUtils
+
+import nutcore._
+
+import utils._
+import bus.simplebus._
+import difftest._
+
+trait HasBackendConst{
+  // val multiIssue = true
+  val robSize = 16
+  val robWidth = 2
+  val robInstCapacity = robSize * robWidth
+  val checkpointSize = 4 // register map checkpoint size
+  val brTagWidth = log2Up(checkpointSize)
+  val prfAddrWidth = log2Up(robSize) + log2Up(robWidth) // physical rf addr width
+
+  val DispatchWidth = 2
+  val CommitWidth = 2
+  val RetireWidth = 2
+
+  val enableCheckpoint = true
+}
