+++ src/main/scala/nutcore/backend/dynamic/EP.scala
@@ -1,8 +1,11 @@
-package nutcore
+
+package nutcore.backend
 
 import chisel3._
 import chisel3.util._
 import chisel3.util.experimental.BoringUtils
+
+import nutcore._
 
 import utils._
