+++ src/main/scala/nutcore/backend/fu/ALU.scala
@@ -20,6 +20,8 @@ import chisel3._
 import chisel3.util._
 import chisel3.util.experimental.BoringUtils
 
+import nutcore.frontend.instr_fetch.branch_predict._
+
 import utils._
 import difftest._
 import top.Settings
