+++ src/main/scala/nutcore/frontend/instr_align_buffer/NaiveIBF.scala
