+++ src/main/scala/nutcore/frontend/instr_fetch/branch_predict/BranchPredictCommons.scala
@@ -1,0 +1,60 @@
+/**************************************************************************************
+* Copyright (c) 2020 Institute of Computing Technology, CAS
+* Copyright (c) 2020 University of Chinese Academy of Sciences
+* 
+* NutShell is licensed under Mulan PSL v2.
+* You can use this software according to the terms and conditions of the Mulan PSL v2. 
+* You may obtain a copy of Mulan PSL v2 at:
+*             http://license.coscl.org.cn/MulanPSL2 
+* 
+* THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER 
+* EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY OR 
+* FIT FOR A PARTICULAR PURPOSE.  
+*
+* See the Mulan PSL v2 for more details.  
+***************************************************************************************/
+
+package nutcore.frontend.instr_fetch.branch_predict
+
+import chisel3._
+import chisel3.util._
+import chisel3.util.experimental.BoringUtils
+
+import nutcore._
+
+import utils._
+import top.Settings
+
+class TableAddr(val idxBits: Int) extends NutCoreBundle {
+  val padLen = if (Settings.get("IsRV32") || !Settings.get("EnableOutOfOrderExec")) 2 else 3
+  def tagBits = VAddrBits - padLen - idxBits
+
+  //val res = UInt((AddrBits - VAddrBits).W)
+  val tag = UInt(tagBits.W)
+  val idx = UInt(idxBits.W)
+  val pad = UInt(padLen.W)
+
+  def fromUInt(x: UInt) = x.asTypeOf(UInt(VAddrBits.W)).asTypeOf(this)
+  def getTag(x: UInt) = fromUInt(x).tag
+  def getIdx(x: UInt) = fromUInt(x).idx
+}
+
+object BTBtype {
+  def B = "b00".U  // branch
+  def J = "b01".U  // jump
+  def I = "b10".U  // indirect
+  def R = "b11".U  // return
+
+  def apply() = UInt(2.W)
+}
+
+class BPUUpdateReq extends NutCoreBundle {
+  val valid = Output(Bool())
+  val pc = Output(UInt(VAddrBits.W))
+  val isMissPredict = Output(Bool())
+  val actualTarget = Output(UInt(VAddrBits.W))
+  val actualTaken = Output(Bool())  // for branch
+  val fuOpType = Output(FuOpType())
+  val btbType = Output(BTBtype())
+  val isRVC = Output(Bool()) // for ras, save PC+2 to stack if is RVC
+}
