+++ src/main/scala/nutcore/frontend/instr_fetch/branch_predict/Legacy.scala
@@ -1,0 +1,39 @@
+// package nutcore.frontend.instr_fetch.branch_predict
+
+// import chisel3._
+// import chisel3.util._
+// import chisel3.util.experimental.BoringUtils
+
+// import nutcore._
+
+// import utils._
+// import top.Settings
+
+//---- Legacy BPUs ----
+/*
+class BranchPredictLegacy extends NutCoreModule {
+  val io = IO(new Bundle {
+    val in = Flipped(Valid(new CtrlFlowIO))
+    val out = new RedirectIO
+  })
+
+  val instr = io.in.bits.instr
+  val immJ = SignExt(Cat(instr(31), instr(19, 12), instr(20), instr(30, 21), 0.U(1.W)), XLEN)
+  val immB = SignExt(Cat(instr(31), instr(7), instr(30, 25), instr(11, 8), 0.U(1.W)), XLEN)
+  val table = Array(
+    RV32I_BRUInstr.JAL  -> List(immJ, true.B),
+    RV32I_BRUInstr.BNE  -> List(immB, instr(31)),
+    RV32I_BRUInstr.BEQ  -> List(immB, instr(31)),
+    RV32I_BRUInstr.BLT  -> List(immB, instr(31)),
+    RV32I_BRUInstr.BGE  -> List(immB, instr(31)),
+    RV32I_BRUInstr.BLTU -> List(immB, instr(31)),
+    RV32I_BRUInstr.BGEU -> List(immB, instr(31))
+  )
+  val default = List(immB, false.B)
+  val offset :: predict :: Nil = ListLookup(instr, default, table)
+
+  io.out.target := io.in.bits.pc + offset
+  io.out.valid := io.in.valid && predict(0)
+  io.out.rtype := 0.U
+}
+*/
