+++ src/main/scala/nutcore/isa/RVI.scala
@@ -19,6 +19,8 @@ package nutcore
 import chisel3._
 import chisel3.util._
 
+import nutcore.frontend.instr_fetch.branch_predict._
+
 object RV32I_ALUInstr extends HasInstrType with HasNutCoreParameter {
   def ADDI    = BitPat("b????????????_?????_000_?????_0010011")
   def SLLI    = if (XLEN == 32) BitPat("b0000000?????_?????_001_?????_0010011")
