+++ src/main/scala/nutcore/mem/cache/Cache.scala
